# social-media/openapi.yaml
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  /registration:
    post:
      tags:
        - user
      summary: Register a new user
      operationId: user_registration_registration_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseUser'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /update-profile:
    put:
      tags:
        - user
      summary: Update profile data
      operationId: update_profile_update_profile_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseUpdateData'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /authentication:
    get:
      tags:
        - user
      summary: User authorization
      operationId: user_authorization_authentication_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseAuthorizationData'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /create-post:
    post:
      tags:
        - post
      summary: Create new post
      operationId: create_post_create_post_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseNewPost'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewPostResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /update-post:
    put:
      tags:
        - post
      summary: Update post
      operationId: update_post_update_post_put
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseUpdatePost'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /delete-post:
    delete:
      tags:
        - post
      summary: Delete post by its id
      operationId: delete_post_delete_post_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaseDeletePost'
        required: true
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /get-post-by-id:
    get:
      tags:
        - post
      summary: Get post by its id
      operationId: get_post_by_id_get_post_by_id_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetPostById'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostItem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /get-posts:
    get:
      tags:
        - post
      summary: Get all posts for user
      operationId: get_posts_get_posts_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestGetPosts'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/PostItem'
                type: array
                title: Response Get Posts Get Posts Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
    BaseAuthorizationData:
      properties:
        login:
          type: string
          title: Login
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - password
      title: BaseAuthorizationData
    BaseDeletePost:
      properties:
        login:
          type: string
          title: Login
        post_id:
          type: integer
          title: Post Id
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - post_id
        - password
      title: BaseDeletePost
    BaseNewPost:
      properties:
        login:
          type: string
          title: Login
        title:
          type: string
          title: Title
        content:
          type: string
          title: Content
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - title
        - content
        - password
      title: BaseNewPost
    BaseUpdateData:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        date_of_birth:
          type: string
          title: Date Of Birth
        mail:
          type: string
          title: Mail
        phone_number:
          type: string
          title: Phone Number
        login:
          type: string
          title: Login
      type: object
      required:
        - first_name
        - last_name
        - date_of_birth
        - mail
        - phone_number
        - login
      title: BaseUpdateData
    BaseUpdatePost:
      properties:
        login:
          type: string
          title: Login
        post_id:
          type: integer
          title: Post Id
        new_title:
          type: string
          title: New Title
        new_content:
          type: string
          title: New Content
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - post_id
        - new_title
        - new_content
        - password
      title: BaseUpdatePost
    BaseUser:
      properties:
        first_name:
          type: string
          title: First Name
        last_name:
          type: string
          title: Last Name
        date_of_birth:
          type: string
          title: Date Of Birth
        mail:
          type: string
          title: Mail
        phone_number:
          type: string
          title: Phone Number
        login:
          type: string
          title: Login
        password:
          type: string
          title: Password
      type: object
      required:
        - first_name
        - last_name
        - date_of_birth
        - mail
        - phone_number
        - login
        - password
      title: BaseUser
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    NewPostResponse:
      properties:
        post_id:
          type: integer
          title: Post Id
      type: object
      required:
        - post_id
      title: NewPostResponse
    PostItem:
      properties:
        id:
          type: integer
          title: Id
        user_id:
          type: integer
          title: User Id
        title:
          type: string
          title: Title
        time:
          type: string
          format: date-time
          title: Time
        content:
          type: string
          title: Content
      type: object
      required:
        - id
        - user_id
        - title
        - time
        - content
      title: PostItem
    RequestGetPostById:
      properties:
        user_login:
          type: string
          title: User Login
        post_id:
          type: integer
          title: Post Id
        password:
          type: string
          title: Password
      type: object
      required:
        - user_login
        - post_id
        - password
      title: RequestGetPostById
    RequestGetPosts:
      properties:
        login:
          type: string
          title: Login
        password:
          type: string
          title: Password
      type: object
      required:
        - login
        - password
      title: RequestGetPosts
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError